# backend/monad-backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Monad API",
        "description": "Authentication and user management API for Monad application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": ""
        }
    ],
    "paths": {
...
        "/api/storage/config": {
            "get": {
                "tags": [
                    "Storage"
                ],
                "responses": {
                    "200": {
                        "description": "Storage configuration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "maxFileSize": {
                                            "type": "integer",
                                            "example": 10485760,
                                            "description": "Maximum file size in bytes"
                                        },
                                        "maxFileSizeMB": {
                                            "type": "integer",
                                            "example": 10,
                                            "description": "Maximum file size in megabytes"
                                        },
                                        "allowedContentTypes": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "application/octet-stream",
                                                "application/json",
                                                "text/csv",
                                                "text/plain"
                                            ],
                                            "description": "List of allowed MIME types"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "AUTH_006"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Authentication required"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "summary": "Get storage configuration",
                "description": "Returns the storage service configuration including allowed file types and size limits",
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/storage/test": {
            "get": {
                "tags": [
                    "Storage"
                ],
                "responses": {
                    "200": {
                        "description": "Connection test result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "bucket": {
                                            "type": "string",
                                            "example": "my-bucket"
                                        },
                                        "region": {
                                            "type": "string",
                                            "example": "eu-central-1"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully connected to S3 bucket"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "AUTH_006"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Authentication required"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "summary": "Test S3 connection",
                "description": "Tests the connection to the configured S3 bucket. Use this to verify your S3 configuration is correct.",
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/storage/upload": {
            "post": {
                "tags": [
                    "Storage"
                ],
                "responses": {
                    "200": {
                        "description": "File uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "objectKey": {
                                            "type": "string",
                                            "example": "uploads/550e8400-e29b-41d4-a716-446655440000/a1b2c3d4/ble_data.csv",
                                            "description": "S3 object key where the file was stored"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "https://bucket.s3.region.amazonaws.com/uploads/...",
                                            "description": "URL of the uploaded file"
                                        },
                                        "size": {
                                            "type": "integer",
                                            "example": 1048576,
                                            "description": "File size in bytes"
                                        },
                                        "contentType": {
                                            "type": "string",
                                            "example": "text/csv",
                                            "description": "MIME type of the uploaded file"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "STORAGE_301"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "File size exceeds maximum allowed (10 MB)"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "AUTH_006"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Authentication required"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Upload failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "STORAGE_300"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "File upload failed"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "summary": "Upload file to S3 (direct stream)",
                "description": "Uploads a file directly to S3 by streaming from request body. NO temp file is created on the backend. Send raw binary body with X-Filename header. Max file size: 50 MB.",
                "parameters": [
                    {
                        "name": "X-Filename",
                        "in": "header",
                        "description": "Original filename (required)",
                        "required": true,
                        "deprecated": false,
                        "schema": {
                            "type": "string",
                            "example": "ble_data.csv"
                        },
                        "style": "simple",
                        "explode": false
                    }
                ],
                "requestBody": {
                    "description": "Raw binary file content (NOT multipart/form-data)",
                    "content": {
                        "application/octet-stream": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "JWT": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [],
...
}